[fea] add mapper to classes USER,Category for conversion

+++ src/main/java/com/example/hinking/models/User.java
@@ -21,7 +21,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userID;
+    private Long userID;
 
     private String name;
